requirements.txt

python-dotenv
pymongo
pandas